Factor out deriveGraph
